Updated so it can show that FPS is be lower than 10

+++ source/mainwindow.h
@@ -97,7 +97,7 @@ private:
     QTimer *stopwatchtimer;
 
     //FPS
-    int recordingFPS;
+    double recordingFPS;
 
 
 protected:
